Give authorization to parent to pick up children. Give access to mobile app to emails

+++ app/main/crud/owner_crud.py
@@ -213,60 +213,33 @@ class CRUDOwner(CRUDBase[models.Owner, schemas.AdministratorCreate, schemas.Admi
         send_account_creation_email(email_to=obj_in.email, prefered_language=lang, name=obj_in.firstname,
                                     password=password, login_link=Config.LOGIN_LINK.format(lang))
         return user
-
     
     @classmethod
-    def give_parent_pickup_child_authorization_for_nursery(cls, db: Session, obj_in: schemas.ChildrenConfirmation, added_by: models.Owner):
-        
-        parent = crud.parent.get_by_email(db, obj_in.parent_email)
+    def give_parent_pickup_child_authorization_for_nursery(cls, db: Session, parent_guest: models.ParentGuest):
+        parent_guest.has_pickup_child_authorization = not parent_guest.has_pickup_child_authorization
+        db.commit()
+        return parent_guest
 
-        exist_pickup_parent_child_for_nursery  = db.query(models.PickUpParentChild).\
-            filter(models.PickUpParentChild.parent_email == obj_in.parent_email).\
-            filter(models.PickUpParentChild.child_uuid == obj_in.child_uuid).\
-            filter(models.PickUpParentChild.nursery_uuid == obj_in.nursery_uuid).\
-            first()
-        
-        if not exist_pickup_parent_child_for_nursery:
-            exist_pickup_parent_child_for_nursery = models.PickUpParentChild(
-                uuid= str(uuid.uuid4()),
-                parent_uuid = parent.uuid,
-                parent_email = obj_in.parent_email,
-                nursery_uuid = obj_in.nursery_uuid,
-                child_uuid = obj_in.child_uuid,
-                added_by_uuid = added_by.uuid
-            )
-            db.add(exist_pickup_parent_child_for_nursery)
-            db.commit()
-            db.refresh(exist_pickup_parent_child_for_nursery)
-    
-        return exist_pickup_parent_child_for_nursery
-    
     @classmethod
-    def confirm_apps_authorization(cls, db: Session, obj_in: schemas.ChildrenConfirmation, added_by: models.Owner,preregistration:models.PreRegistration):
-        
-        is_parent_guest = False
-        parent = db.query(models.Parent).filter(models.Parent.email.ilike(obj_in.parent_email)).first()
-        if not parent:
-            is_parent_guest = True
-            parent = db.query(models.ParentGuest).filter(models.ParentGuest.email.ilike(obj_in.parent_email)).first()
+    def confirm_apps_authorization(cls, db: Session, obj_in: schemas.ChildrenConfirmation, added_by: models.Owner):
 
-        parent_child  = db.query(models.ParentChild).\
+        parent_child = db.query(models.ParentChild).\
             filter(models.ParentChild.parent_email == obj_in.parent_email).\
             filter(models.ParentChild.child_uuid == obj_in.child_uuid).\
             filter(models.ParentChild.nursery_uuid == obj_in.nursery_uuid).\
             first()
         
         if obj_in.status in ['REFUSED']:
-            db.delete(parent_child)
-            db.commit()
-            parent_child = None
+            if parent_child:
+                db.delete(parent_child)
+                db.commit()
+                parent_child = None
         
         else:
-        
             if not parent_child:
                 parent_child = models.ParentChild(
                     uuid= str(uuid.uuid4()),
-                    parent_uuid = parent.uuid,
+                    # parent_uuid = parent.uuid,
                     parent_email = obj_in.parent_email,
                     nursery_uuid = obj_in.nursery_uuid,
                     child_uuid = obj_in.child_uuid,
@@ -274,60 +247,11 @@ class CRUDOwner(CRUDBase[models.Owner, schemas.AdministratorCreate, schemas.Admi
                 )
                 db.add(parent_child)
                 db.flush()
-                
-            if is_parent_guest == True:
-                contract = crud.contract.get_contract_by_uuid(db=db, uuid=preregistration.contract_uuid)
-                if contract:
-                    parent = db.query(models.ParentGuest).filter(models.ParentGuest.uuid == parent_child.parent_uuid).first()
-                    if parent:
-                        if contract not in parent.contracts:
-                            parent.contracts.append(contract)
-
-                # To check this TODO
-                # code = generate_code(length=12)
-                # code= str(code[0:6]) 
-                
-                parent.is_new_user = True
-                
-                # user_code: models.ParentActionValidation = db.query(models.ParentActionValidation).filter(
-                # models.ParentActionValidation.user_uuid == parent.uuid)
-
-                # if user_code.count() > 0:
-                #     user_code.delete()
-
-                # # print("user_code1:")
-                # db_code = models.ParentActionValidation(
-                #     uuid=str(uuid.uuid4()),
-                #     code=code,
-                #     user_uuid=parent.uuid,
-                #     value=code,
-                #     expired_date=datetime.now() + timedelta(minutes=30)
-                # )
-
-                # db.add(db_code)
-                # db.commit()
-
-                role = crud.role.get_by_code(db=db, code="parent")
-                if not role:
-                    raise HTTPException(status_code=404, detail=__("role-not-found"))
-                    
-                # Generate password and send to user : username and password on the email
-                password = generate_password()
-                print("Parent Guest: ", password)
-                parent.password_hash = get_password_hash(password)
-
-                parent.role_uuid = role.uuid,
-                parent.status = models.UserStatusType.ACTIVED
-                db.commit()
-                db.refresh(parent)
-
-                # send_account_confirmation_email(email_to=parent.email, name=(parent.firstname+parent.lastname), token=code, valid_minutes=30)
 
             db.commit()
             db.refresh(parent_child)
 
         return parent_child
     
+
 owner = CRUDOwner(models.Owner)
-
-
